app/requirements.txt
django
django_restframework_apiview
uwsgi
